Renames directory cate to cate_plugin. Conflicted with cate library in dev mode.
